requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
torchxrayvision>=1.1.0
grad-cam
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
scipy>=1.10.0
pyyaml>=6.0
tqdm>=4.65.0
captum>=0.6.0
tensorboard>=2.13.0
